requirements.txt
playwright==1.40.0
gologin==2025.7.3
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
requests==2.31.0
selenium==4.15.2
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
python-telegram-bot==20.7
beautifulsoup4==4.12.2
lxml==4.9.3
asyncio==3.4.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
pillow==10.0.1
psutil==5.9.6
cryptography==41.0.7
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0